mount the SSH_AUTH_SOCK socket for ssh-agent
If DEPENDABOT_SSH_CREDS is non-empty, mount the SSH_AUTH_SOCK
into the docker-dev-shell container to propagate the SSH agent
credentials and allow git access to servers/repositories
which require the ssh credentials (which would be especially
important in corporate environments).

The added volume mapping in 'docker run' applies only if
both ${DEPENDABOT_SSH_CREDS} and ${SSH_AUTH_SOCK} are not empty.
For Mac with Docker or Rancher Desktop, the following
value can be used:
DEPENDABOT_SSH_CREDS=1 SSH_AUTH_SOCK=/run/host-services/ssh-auth.sock \
  ./bin/docker-dev-shell go_modules -r

PR redoes with the addition of DEPENDABOT_SSH_CREDS flag the change
merged in #11683 but reverted in #11818.
Addresses in part issue #11544.

Signed-off-by: Dmitry Savintsev <dsavints@gmail.com>